chore: move desktop integration tests to desktop folder
