+++ frontend/appflowy_flutter/integration_test/desktop/board/board_group_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:appflowy_board/appflowy_board.dart';
 
-import '../util/util.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
