+++ frontend/appflowy_flutter/integration_test/desktop/board/board_row_test.dart
@@ -7,8 +7,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/util.dart';
-import '../util/database_test_op.dart';
+import '../../util/util.dart';
+import '../../util/database_test_op.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
