+++ frontend/appflowy_flutter/integration_test/desktop/cloud/anon_user_continue_test.dart
@@ -16,9 +16,9 @@ import 'package:flowy_infra/uuid.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path/path.dart' as p;
 import 'package:integration_test/integration_test.dart';
-import '../util/dir.dart';
-import '../util/mock/mock_file_picker.dart';
-import '../util/util.dart';
+import '../../util/dir.dart';
+import '../../util/mock/mock_file_picker.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
