+++ frontend/appflowy_flutter/integration_test/desktop/cloud/appflowy_cloud_auth_test.dart
@@ -11,10 +11,11 @@ import 'package:appflowy/workspace/presentation/settings/widgets/settings_user_v
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra/uuid.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:path/path.dart' as p;
 import 'package:integration_test/integration_test.dart';
-import '../util/mock/mock_file_picker.dart';
-import '../util/util.dart';
+import 'package:path/path.dart' as p;
+
+import '../../util/mock/mock_file_picker.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
@@ -47,13 +48,13 @@ void main() {
       tester.expectToSeeGoogleLoginButton();
     });
 
-    testWidgets('sign in as annoymous', (tester) async {
+    testWidgets('sign in as anonymous', (tester) async {
       await tester.initializeAppFlowy(
         cloudType: AuthenticatorType.appflowyCloudSelfHost,
       );
       await tester.tapSignInAsGuest();
 
-      // should not see the sync setting page when sign in as annoymous
+      // should not see the sync setting page when sign in as anonymous
       await tester.openSettings();
       await tester.openSettingsPage(SettingsPage.user);
       tester.expectToSeeGoogleLoginButton();
