+++ frontend/appflowy_flutter/integration_test/desktop/cloud/supabase_auth_test.dart
@@ -6,7 +6,8 @@ import 'package:appflowy/workspace/presentation/settings/widgets/settings_user_v
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import '../util/util.dart';
+
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
@@ -35,11 +36,11 @@ void main() {
       tester.expectToSeeGoogleLoginButton();
     });
 
-    testWidgets('sign in as annoymous', (tester) async {
+    testWidgets('sign in as anonymous', (tester) async {
       await tester.initializeAppFlowy(cloudType: AuthenticatorType.supabase);
       await tester.tapSignInAsGuest();
 
-      // should not see the sync setting page when sign in as annoymous
+      // should not see the sync setting page when sign in as anonymous
       await tester.openSettings();
       await tester.openSettingsPage(SettingsPage.user);
       tester.expectToSeeGoogleLoginButton();
