+++ frontend/appflowy_flutter/integration_test/desktop/cloud/user_setting_sync_test.dart
@@ -18,11 +18,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path/path.dart' as p;
 import 'package:integration_test/integration_test.dart';
-import '../util/database_test_op.dart';
-import '../util/dir.dart';
-import '../util/emoji.dart';
-import '../util/mock/mock_file_picker.dart';
-import '../util/util.dart';
+import '../../util/database_test_op.dart';
+import '../../util/dir.dart';
+import '../../util/emoji.dart';
+import '../../util/mock/mock_file_picker.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
