+++ frontend/appflowy_flutter/integration_test/desktop/database/database_row_page_test.dart
@@ -7,9 +7,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/database_test_op.dart';
-import '../util/emoji.dart';
-import '../util/util.dart';
+import '../../util/database_test_op.dart';
+import '../../util/emoji.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
