+++ frontend/appflowy_flutter/integration_test/desktop/database/database_setting_test.dart
@@ -3,8 +3,8 @@ import 'package:appflowy_backend/protobuf/flowy-folder/view.pb.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/database_test_op.dart';
-import '../util/util.dart';
+import '../../util/database_test_op.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
