+++ frontend/appflowy_flutter/integration_test/desktop/database/database_share_test.dart
@@ -2,7 +2,7 @@ import 'package:appflowy_backend/protobuf/flowy-database2/field_entities.pbenum.
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/database_test_op.dart';
+import '../../util/database_test_op.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
