+++ frontend/appflowy_flutter/integration_test/desktop/document/document_alignment_test.dart
@@ -5,8 +5,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/keyboard.dart';
-import '../util/util.dart';
+import '../../util/keyboard.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
