+++ frontend/appflowy_flutter/integration_test/desktop/document/document_with_cover_image_test.dart
@@ -7,8 +7,8 @@ import 'package:flutter_emoji_mart/flutter_emoji_mart.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/emoji.dart';
-import '../util/util.dart';
+import '../../util/emoji.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
