+++ frontend/appflowy_flutter/integration_test/desktop/document/document_with_inline_page_test.dart
@@ -5,7 +5,7 @@ import 'package:flowy_infra_ui/widget/error_page.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/util.dart';
+import '../../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
