+++ frontend/appflowy_flutter/integration_test/desktop/empty_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import 'util/util.dart';
+import '../util/util.dart';
 
 // This test is meaningless, just for preventing the CI from failing.
 void main() {
