+++ frontend/appflowy_flutter/integration_test/desktop/import_files_test.dart
@@ -3,10 +3,10 @@ import 'dart:io';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-
-import 'util/mock/mock_file_picker.dart';
-import 'util/util.dart';
 import 'package:path/path.dart' as p;
+
+import '../util/mock/mock_file_picker.dart';
+import '../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
