+++ frontend/appflowy_flutter/integration_test/desktop/reminder/document_reminder_test.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/services.dart';
-
 import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/mention/mention_block.dart';
 import 'package:appflowy/user/application/user_settings_service.dart';
@@ -11,14 +9,15 @@ import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:calendar_view/calendar_view.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra_ui/style_widget/text_field.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/base.dart';
-import '../util/common_operations.dart';
-import '../util/editor_test_operations.dart';
-import '../util/expectation.dart';
-import '../util/keyboard.dart';
+import '../../util/base.dart';
+import '../../util/common_operations.dart';
+import '../../util/editor_test_operations.dart';
+import '../../util/expectation.dart';
+import '../../util/keyboard.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
