+++ frontend/appflowy_flutter/integration_test/desktop/share_markdown_test.dart
@@ -5,8 +5,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:path/path.dart' as p;
 
-import 'util/mock/mock_file_picker.dart';
-import 'util/util.dart';
+import '../util/mock/mock_file_picker.dart';
+import '../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
@@ -104,12 +104,12 @@ fn main() {
 > Click `?` at the bottom right for help and support.
 
 > 🥰
-> 
+>
 > Like AppFlowy? Follow us:
 > [GitHub](https://github.com/AppFlowy-IO/AppFlowy)
 > [Twitter](https://twitter.com/appflowy): @appflowy
 > [Newsletter](https://blog-appflowy.ghost.io/)
-> 
+>
