+++ frontend/appflowy_flutter/integration_test/desktop/sidebar/sidebar_icon_test.dart
@@ -2,9 +2,9 @@ import 'package:appflowy_backend/protobuf/flowy-folder/view.pb.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import '../util/base.dart';
-import '../util/common_operations.dart';
-import '../util/expectation.dart';
+import '../../util/base.dart';
+import '../../util/common_operations.dart';
+import '../../util/expectation.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
