+++ frontend/appflowy_flutter/integration_test/desktop/switch_folder_test.dart
@@ -7,8 +7,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:path/path.dart' as p;
 
-import 'util/mock/mock_file_picker.dart';
-import 'util/util.dart';
+import '../util/mock/mock_file_picker.dart';
+import '../util/util.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
